moved integration tests to separate module and create unified coverage report
