+++ .travis.yml
@@ -1,3 +1,4 @@
 language: java
-
-script: "mvn clean install"
+cache:
+  directories:
+    - $HOME/.m2
