+++ spring-cloud-config-aws-kms-tests/spring-cloud-config-aws-kms-integration-test/src/test/java/de/zalando/spring/cloud/config/aws/kms/it/KmsRegionConfigurationTest.java
@@ -1,4 +1,4 @@
-package de.zalando.spring.cloud.config.aws.kms;
+package de.zalando.spring.cloud.config.aws.kms.it;
 
 import com.amazonaws.services.kms.AWSKMS;
 import com.amazonaws.services.kms.AWSKMSClient;
@@ -17,20 +17,21 @@ import org.springframework.util.ReflectionUtils;
 
 import java.lang.reflect.Field;
 import java.net.URI;
+import java.util.Objects;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNotNull;
+import static org.junit.Assert.assertNull;
 import static org.junit.Assert.assertTrue;
 
 /**
- * During this integration test, a real AWSKMSClient is created, but there are no encrypted properties, so te client is
+ * During this integration test, a real AWSKMSClient is created, but there are no encrypted properties, so the client is
  * never used.<br/>
  * See src/test/resources/*-noEncryption.yml files.
  */
 @SpringBootTest
-@ActiveProfiles("noEncryptionEndpoint")
-public class NoKmsEncryptionIntegrationEndpointConfigurationTest {
+@ActiveProfiles("noEncryption")
+public class KmsRegionConfigurationTest {
 
     @ClassRule
     public static final SpringClassRule SPRING_CLASS_RULE = new SpringClassRule();
@@ -44,7 +45,7 @@ public class NoKmsEncryptionIntegrationEndpointConfigurationTest {
     private AWSKMS kms;
 
     @Test
-    public void testPropertyHasBeenDecrypted() throws Exception {
+    public void testPropertyIsAvailable() {
         assertThat(secret).isEqualTo("secret");
     }
 
@@ -53,29 +54,20 @@ public class NoKmsEncryptionIntegrationEndpointConfigurationTest {
         assertNotNull(kms);
         assertTrue(kms instanceof AWSKMSClient);
 
+        // endpoint configured based on aws.kms.region property
         AWSKMSClient client = (AWSKMSClient) kms;
-
-        // prove aws.kms.endpoint.service-endpoint was used to configure the kms client
-        Field endpointField = ReflectionUtils.findField(AWSKMSClient.class, "endpoint");
-        ReflectionUtils.makeAccessible(endpointField);
-        Object endpointObject = ReflectionUtils.getField(endpointField, client);
+        Field field = ReflectionUtils.findField(AWSKMSClient.class, "endpoint");
+        ReflectionUtils.makeAccessible(Objects.requireNonNull(field));
+        Object endpointObject = ReflectionUtils.getField(field, client);
         assertNotNull(endpointObject);
         assertTrue(endpointObject instanceof URI);
         URI endpoint = (URI) endpointObject;
-        assertTrue(endpoint.toString().contains("us-east-1"));
+        assertTrue(endpoint.toString().contains("eu-central-1"));
 
-        // prove override was issued via the aws.kms.endpoint.signing-region property
+        // no override should occur in this configuration
         Field signerRegionField = ReflectionUtils.findField(AWSKMSClient.class, "signerRegionOverride");
-        ReflectionUtils.makeAccessible(signerRegionField);
+        ReflectionUtils.makeAccessible(Objects.requireNonNull(signerRegionField));
         Object signerRegionObject = ReflectionUtils.getField(signerRegionField, client);
-        assertNotNull(signerRegionObject);
-        assertTrue(signerRegionObject instanceof String);
-        String signerRegion = (String) signerRegionObject;
-        assertEquals("us-east-2", signerRegion);
-    }
-
-    @Configuration
-    @EnableAutoConfiguration
-    static class TestConfig {
+        assertNull(signerRegionObject);
     }
 }
