+++ spring-cloud-config-aws-kms-tests/spring-cloud-config-aws-kms-integration-test/src/test/java/de/zalando/spring/cloud/config/aws/kms/it/TestApplication.java
@@ -1,0 +1,7 @@
+package de.zalando.spring.cloud.config.aws.kms.it;
+
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class TestApplication {
+}
