+++ spring-cloud-config-aws-kms-tests/spring-cloud-config-aws-kms-integration-test/src/test/resources/application-encryption.yml
