+++ spring-cloud-config-aws-kms-tests/spring-cloud-config-aws-kms-integration-test/src/test/resources/bootstrap-encryption.yml
